Add App link to the Deployment details pane
