+++ .changeset/lemon-buckets-look.md
@@ -1,0 +1,5 @@
+---
+'@giantswarm/backstage-plugin-gs': minor
+---
+
+Added catalog entity link to the Deployment details pane.
