+++ plugins/gs/src/components/deployments/AppDetails/AppDetails.tsx
@@ -2,6 +2,7 @@ import React, { ReactNode } from 'react';
 import { Link as RouterLink } from 'react-router-dom';
 import { EmptyState, Progress, WarningPanel } from '@backstage/core-components';
 import { useRouteRef } from '@backstage/core-plugin-api';
+import { EntityRefLink } from '@backstage/plugin-catalog-react';
 import {
   Box,
   Card,
@@ -22,7 +23,7 @@ import {
   App,
   AppKind,
 } from '@giantswarm/backstage-plugin-gs-common';
-import { useResource } from '../../hooks';
+import { useCatalogEntitiesForDeployments, useResource } from '../../hooks';
 import {
   formatAppCatalogName,
   formatSource,
@@ -33,6 +34,7 @@ import {
   DateComponent,
   GrafanaDashboardLink,
   Heading,
+  NotAvailable,
   StructuredMetadataList,
 } from '../../UI';
 import { AppDetailsStatus } from '../AppDetailsStatus';
@@ -69,6 +71,8 @@ export const AppDetails = ({
 
   const clusterRouteLink = useRouteRef(clusterDetailsRouteRef);
 
+  const catalogEntitiesMap = useCatalogEntitiesForDeployments();
+
   if (isLoading) {
     return <Progress />;
   }
@@ -101,7 +105,7 @@ export const AppDetails = ({
   const sourceName = formatAppCatalogName(getAppCatalogName(app) ?? '');
   const chartName = getAppChartName(app);
 
-  let clusterEl: ReactNode = clusterName ? clusterName : 'n/a';
+  let clusterEl: ReactNode = clusterName ? clusterName : <NotAvailable />;
   if (clusterName && clusterNamespace) {
     clusterEl = (
       <Link
@@ -116,6 +120,13 @@ export const AppDetails = ({
       </Link>
     );
   }
+
+  const entityRef = chartName ? catalogEntitiesMap[chartName] : undefined;
+  const entityLink: ReactNode = entityRef ? (
+    <EntityRefLink entityRef={entityRef} />
+  ) : (
+    <NotAvailable />
+  );
 
   return (
     <div>
@@ -146,6 +157,7 @@ export const AppDetails = ({
                 metadata={{
                   Installation: installationName,
                   Cluster: clusterEl,
+                  App: entityLink,
                 }}
               />
             </CardContent>
